default to max_k=31

+++ include/kmer.hpp
@@ -2,7 +2,7 @@
 
 namespace sshash {
 
-// typedef uint64_t kmer_t;
-typedef __uint128_t kmer_t;
+typedef uint64_t kmer_t;
+// typedef __uint128_t kmer_t;
 
 }  // namespace sshash
